Add CI and make tests support (#215)
* Add mocking to lambda_trigger tests

Tests were currently unable to pass because they were live tests. Here we've
added mocked responses from github and releases. A make target was added to
document how the raw files were obtained and formatted.

Minor fix to the way error checks were handled so that we could provide better
diagnostic information when there was an error.

* Correct typo in gitignore

* Add additional makefiles for the rest

* Add CI build step

We limit the builds to pull requests and any branches that touch the utilities

Update lambda_trigger go version