Ignore variant releases overridding core versions
This addresses the root cause behind mistakenly publishing vault's fips
variants as the main version in #223, #206, and most recently, 64171e0a.

The Releases API does not allow us to filter by variant, so we hardcode the
variant we are interested in. This restricts publishing to just our base
variants of enterprise and OSS. Perhaps in the future we can revisit this to
support homebrew publishing of our variants.