lambda_trigger: Support new products (#221)
* Add support for new products

If we 404 while checking for an existing Formula/Cask, that's our signal this is a new product.

* Add tests for new product support and lambda handling

Testing the lambda event processing in a bit messy since it's effectively an end-to-end test and requires mocking a few different endpoints (GH raw, GH API, and Releases). But I figure it'd probably be a good idea to validate our event processing since errors there are hidden in cloudwatch.